+++ src/types/puzzle.ts
@@ -21,6 +21,7 @@ export interface Fen {
     current: string;
     previous: string;
 }
+
 export interface PuzzleIndex {
   x: number;
   y: number;
