+++ src/utils/attemptMove.ts
@@ -1,19 +1,3 @@
 import { Chess } from "chess.js";
 
-export const attemptMove = (
-  game: Chess,
-  sourceSquare: string,
-  targetSquare: string
-) => {
-  try {
-    const move = game.move({
-      from: sourceSquare,
-      to: targetSquare,
-      promotion: "q",
-    });
 
-    return move;
-  } catch (error) {
-    console.error(error);
-  }
-};
