+++ src/utils/calculateBoardSize.ts
@@ -1,13 +1,15 @@
 import { BOARD_DIMENSIONS } from "../constants";
 
-
 export const calculateBoardSize = (
-windowWidth: number, windowHeight: number, offset: number = 100  ): number => {
-    return Math.max(
-      BOARD_DIMENSIONS.MIN_SIZE,
-      Math.min(
-        BOARD_DIMENSIONS.MAX_SIZE,
-        Math.min(windowWidth - offset, windowHeight - offset)
-      )
-    );
-  };
+  windowWidth: number,
+  windowHeight: number,
+  offset: number = 100
+): number => {
+  return Math.max(
+    BOARD_DIMENSIONS.MIN_SIZE,
+    Math.min(
+      BOARD_DIMENSIONS.MAX_SIZE,
+      Math.min(windowWidth - offset, windowHeight - offset)
+    )
+  );
+};
