--- src/utils/chess/checkGoodMove.ts
@@ -1,7 +1,0 @@
-import { normalizeCastlingMove } from "./normalizeCastle";
-
-const checkGoodMove = (acceptableMoves: string[], move: string) => {
-  return acceptableMoves.includes(normalizeCastlingMove(move as string));
-};
-
-export default checkGoodMove;
